gestion de paiement et optimisation des endpoints

+++ server/resolvers/mutation/entreprise/delete.js
@@ -2,19 +2,28 @@ const client = require('../../../services/connection')
 
 
 module.exports = {
-        delEtp: (root,args) => {
+        delEtp: (root,args,context) => {
         try{
-            return new Promise((resolve,reject) => {
-                client.query('DELETE FROM "Entreprises" WHERE ("id" = $1) RETURNING *',[args.id],function(err,result){
-                    if (err){
-                        console.log(err);
-                        reject('Id of entreprise is undefined')
+            if (!(context.userId)){
+                return new GraphQLError('token invalid',{
+                    extensions:{
+                        code:"token invalide"
                     }
-                    else{
-                       resolve(result.rows[0])
-                    }
-                })
             })
+            }
+            else{
+                return new Promise((resolve,reject) => {
+                    client.query('DELETE FROM "Entreprises" WHERE ("id" = $1) RETURNING *',[args.id],function(err,result){
+                        if (err){
+                            console.log(err);
+                            reject('Id of entreprise is undefined')
+                        }
+                        else{
+                           resolve(result.rows[0])
+                        }
+                    })
+                })
+            }
         }
         catch(err){
             console.log(err)
