+++ server/resolvers/mutation/publicités/delete.js
@@ -1,18 +1,27 @@
 const client = require('../../../services/connection')
 
 module.exports = {
-    delPublicites:(parent,args) =>{
+    delPublicites:(parent,args,context) =>{
         try{
-            return new Promise((resolve,reject) =>{
-                client.query('DELETE FROM "Publicités" WHERE (id=$1) RETURNING *',[args.id],function(err,result){
-                    if(err){
-                        reject(err)
+            if (!(context.userId)){
+                return new GraphQLError('token invalid',{
+                    extensions:{
+                        code:"token invalide"
                     }
-                    else{
-                        resolve(result.rows[0])
-                    }
-                })
             })
+            }
+            else{
+                return new Promise((resolve,reject) =>{
+                    client.query('DELETE FROM "Publicités" WHERE (id=$1) RETURNING *',[args.id],function(err,result){
+                        if(err){
+                            reject(err)
+                        }
+                        else{
+                            resolve(result.rows[0])
+                        }
+                    })
+                })
+            }
         }
         catch(err){
             console.log(err)
